refactor: remove rpc
