+++ zero_bin/common/src/bin/rpc/native/state.rs
