--- zero_bin/rpc/Cargo.toml
@@ -1,55 +1,0 @@
-[package]
-name = "rpc"
-authors = ["Polygon Zero <zbrown@polygon.technology>"]
-version = "0.1.0"
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-keywords.workspace = true
-categories.workspace = true
-build = "../common/build.rs"
-
-[dependencies]
-__compat_primitive_types = { workspace = true }
-alloy = { workspace = true }
-alloy-compat = "0.1.0"
-anyhow = { workspace = true }
-clap = { workspace = true }
-futures = { workspace = true }
-hex = { workspace = true }
-itertools = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tokio = { workspace = true }
-tower = { workspace = true, features = ["retry"] }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-url = { workspace = true }
-
-# Local dependencies
-compat = { workspace = true }
-evm_arithmetization = { workspace = true }
-mpt_trie = { workspace = true }
-proof_gen = { workspace = true }
-trace_decoder = { workspace = true }
-zero_bin_common = { workspace = true }
-zk_evm_common = { workspace = true }
-
-[build-dependencies]
-anyhow = { workspace = true }
-cargo_metadata = { workspace = true }
-vergen = { workspace = true }
-
-[features]
-default = ["eth_mainnet"]
-eth_mainnet = [
-    "evm_arithmetization/eth_mainnet",
-    "proof_gen/eth_mainnet",
-    "zero_bin_common/eth_mainnet",
-]
-cdk_erigon = [
-    "evm_arithmetization/cdk_erigon",
-    "proof_gen/cdk_erigon",
-    "zero_bin_common/cdk_erigon",
-]
-polygon_pos = ["evm_arithmetization/polygon_pos", "zero_bin_common/polygon_pos"]
